backend/requirements.txt
Flask==3.0.3
requests==2.31.0
pymongo==4.7.0
python-dotenv==1.0.1
Flask-CORS==4.0.1